+++ include/ir.h
@@ -14,7 +14,7 @@
 #include "pc.h"
 #include "utils.h"
 
-void initialize_ir(size_t instruction_count);
+void initialize_ir(unsigned instruction_count);
 void register_instruction(Opcode opcode, Immediate immediate);
 void free_ir(void);
 void print_ir(void);
